+++ src/main/java/com/gtnewhorizon/cropsnh/crops/vanilla/CropCactus.java
@@ -57,7 +57,7 @@ public class CropCactus extends NHCropCard {
     }
 
     @Override
-    public int getEntityDamage() {
-        return 1;
+    public float getEntityDamage() {
+        return 1.0F;
     }
 }
